feat(frontend): allow action prop in SettingsCard and use for catalogue and changelog card
